refactored firebase auth for modular approach
