+++ index.html
@@ -241,7 +241,9 @@
 		
 		<script type="text/javascript" src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
 		<script src="https://apis.google.com/js/platform.js" async defer></script>
-		<script src="res/scripts/login.js" defer type="module"></script>
+		<script src="res/scripts/firebase/firebase-config.js" type="module"></script>
+		<!-- <script src="res/scripts/firebase/auth.js" defer type="module"></script> -->
+		<script src="res/scripts/main.js" defer type="module"></script>
 		<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
 	</body>
 </html>
