+++ res/styles/general.css
@@ -19,9 +19,4 @@
     background: #555;
   }
 
-  /* Vertical Overflow styling */
-  .overflow-height {
-    box-shadow: 
-    inset 0 8px 8px -8px rgba(0, 0, 0, 0.3),  /* Top shadow */
-    inset 0 -8px 8px -8px rgba(0, 0, 0, 0.3); /* Bottom shadow */
-  }
+
