+++ workplace.html
@@ -100,7 +100,7 @@
                     <h3>My Projects</h3>
                 </div>
                 <div class="project-container overflow-y-auto" id="listContainer">
-                    <div class="d-flex justify-content-start flex-wrap bg-dark">
+                    <div class="d-flex justify-content-start flex-wrap">
                         <!-- Cards for each project created -->
                         <div class="card mx-1 my-1">
                             <div class="card-body project-card-content">
